backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.0
fpdf2==2.8.1
